--- src/components/molecules/CalendarRibbonItem/index.tsx
@@ -1,8 +1,0 @@
-export default function CalendarRibbonItem() {
-  return (
-    <div>
-      <></>
-      <></>
-    </div>
-  );
-}
